tools/wayca-memory-bench: Fix potential overflow issue
In common cases the measuring overhead should be safe to store in an
integer, however since the measure_execute_time() returns a value of
long int it's sane to use the same type to avoid potential overflow
problem.

Fixes: 536d1b5effea ("tools/wayca-memory-bench: fix wayca-memory-bench error")

+++ tools/wayca-memory-bench/wayca-memory-bench.c
@@ -448,7 +448,7 @@ static void walk_circular_list(struct buffer_info *info)
 static void *bench_thread(void *data)
 {
 	struct thread_info *info = data;
-	int overhead;
+	long int overhead;
 
 	if (info->cpu >= 0) {
 		cpu_set_t *cpuset;
